fix(mavlink_receiver): static cast to the time_delay_us

+++ src/modules/mavlink/mavlink_receiver.cpp
@@ -3081,7 +3081,7 @@ MavlinkReceiver::handle_message_local_position_ned_cov(mavlink_message_t *msg)
 	radar_detection.vy = local_position_ned_cov.vy;
 	radar_detection.vz = local_position_ned_cov.vz;
 
-	radar_detection.time_delay_us = hrt_absolute_time() - local_position_ned_cov.time_usec;
+	radar_detection.time_delay_us = static_cast<float>(hrt_absolute_time() - local_position_ned_cov.time_usec);
 
 	radar_detection.pos_cov_x = local_position_ned_cov.covariance[0];
 	radar_detection.pos_cov_y = local_position_ned_cov.covariance[4];
